Fix the strange Multi page problem Jquery mobile and Rails. See https://demos.jquerymobile.com/1.2.0/docs/pages/page-links.html Added data: { ajax: false}  to the edit_all form_tag  Added rel: 'external' attribute to the link_to in the blood_selection list view.
 Both are referring to a Jquery mobile multi page view and THEN

I t's important to note that if you are linking from a mobile page that was loaded via Ajax to a page that contains multiple internal pages, you need to add a rel="external" or data-ajax="false" to the link. This tells the framework to do a full page reload to clear out the Ajax hash in the URL. This is critical because Ajax pages use the hash (#) to track the Ajax history, while multiple internal pages use the hash to indicate internal pages so there will be conflicts in the hash between these two modes. !!!!